requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
aiosmtplib==3.0.1
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.3
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.8
Django==5.0.6
django-ssh-deployer==0.6.2
djangorestframework==3.15.2
drf-yasg==1.21.7
frozenlist==1.4.1
idna==3.7
inflection==0.5.1
kombu==5.3.7
multidict==6.0.5
packaging==24.1
paramiko==3.4.0
prompt_toolkit==3.0.47
pycparser==2.22
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
